requirements.txt
# Core Telegram Bot Dependencies
python-telegram-bot==20.7
telethon==1.34.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Async & Networking
aiohttp==3.9.1
asyncio-mqtt==0.16.1
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
nest-asyncio==1.5.8

# Scheduling & Background Tasks
apscheduler==3.10.4

# Cryptography & Security
cryptography==41.0.7
pycryptodome==3.19.0
pynacl==1.5.0
pyaes==1.6.1
rsa==4.9

# Configuration & Environment
python-dotenv==1.0.0
python-dateutil==2.8.2

# Utilities & Performance
psutil==5.9.6
colorlog==6.8.0
phonenumbers==8.13.26
Pillow==10.1.0
qrcode==7.4.2

# Performance (Unix only)
uvloop==0.19.0; sys_platform != "win32"

# Web Framework (for hosting)
flask==3.0.0
gunicorn==21.2.0

# Database Drivers (optional)
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1